requirements.txt
aiohttp==3.8.5
aiosqlite==0.19.0
beautifulsoup4==4.12.2
redis==4.6.0
python-dateutil==2.8.2
nltk==3.8.1
textstat==0.7.3
lxml==4.9.3
tldextract==3.4.4
elasticsearch==8.9.0
